requirements_kakao.txt
# KakaoTalk OAuth Integration Requirements
django-allauth==0.57.0
requests==2.31.0
PyJWT==2.8.0